refactor(client): collapse admin API layers into adminApi.js
client/src/api/endpoints/admin.js duplicated 11 exports already defined
in adminApi.js with different (uncached vs. cached/adminToken-aware)
behavior, but only endpoints/admin.js was re-exported through the
api/index.js barrel — creating a latent footgun where a future barrel
import could silently resolve to the dead, uncached copy instead of the
live one actually used by every page.

Delete endpoints/admin.js, moving its two non-duplicate exports
(fetchTools, fetchMcpToolCatalog) into adminApi.js and dropping the
unused fetchUsageData (zero importers). Update the small number of
direct/barrel importers of the moved functions to import from adminApi
directly, matching the convention used by the other 81 existing
importers. Add an ESLint no-restricted-imports rule so the deleted path
can't be silently recreated.

Fixes #1743
